+++ Hivemind/Assets/Scripts/Minds/CombatMind.cs
@@ -1,22 +1,22 @@
 ﻿
-    public class CombatMind 
+public class CombatMind
+{
+    private float minEstimetedDifference;
+    private int prefferedHealth;
+
+
+    public CombatMind(float minEstimeted, int prefHealth)
     {
-        private float minEstimetedDifference;
-        private int prefferedHealth;
-
-
-        public CombatMind(float minEstimeted, int prefHealth)
-        {
-            minEstimetedDifference = minEstimeted;
-            prefferedHealth = prefHealth;
-        }
-        public float GetMinEstimetedDifference()
-        {
-            return minEstimetedDifference;
-        }
-
-        public int GetPrefferedHealth()
-        {
-            return prefferedHealth;
-        }
+        minEstimetedDifference = minEstimeted;
+        prefferedHealth = prefHealth;
     }
+    public float GetMinEstimetedDifference()
+    {
+        return minEstimetedDifference;
+    }
+
+    public int GetPrefferedHealth()
+    {
+        return prefferedHealth;
+    }
+}
