+++ Hivemind/Assets/Scripts/Minds/Gathering.cs.meta
