--- Hivemind/Assets/Scripts/Minds/IHasMind.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-    //everything that has this class has minds attachted to it
-    public interface IHasMind
-    {
-        
-    }
