--- Hivemind/Assets/Scripts/Minds/IHasMind.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9f2ae40232696a849b8d96e094d11a79
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
