+++ Hivemind/Assets/Scripts/Minds/IMind.cs
@@ -5,8 +5,8 @@ using System.Text;
 using System.Threading.Tasks;
 
 
-     public interface IMind
-    {
-         void Execute(Ant ant);
-         void Initiate();
-    }
+public interface IMind
+{
+    void Execute(Ant ant);
+    void Initiate();
+}
