+++ Hivemind/Assets/Scripts/Storage.cs
@@ -1,15 +1,15 @@
 ﻿using UnityEngine;
 
 
-    public class Storage : MonoBehaviour
+public class Storage : MonoBehaviour
+{
+    private void Awake()
     {
-        private void Awake()
-        {
-            GameWorld.SetStorage(this);
-        }
-
-        public Vector3 GetPosition()
-        {
-            return gameObject.transform.position;
-        }
+        GameWorld.SetStorage(this);
     }
+
+    public Vector3 GetPosition()
+    {
+        return gameObject.transform.position;
+    }
+}
